src/project/requirements.txt
click==8.1.3
flake8==5.0.0
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
typing_extensions==4.5.0
Werkzeug==2.2.3
loguru==0.6.0